+++ packages/plugins/Fullscreen/src/components/Fullscreen.vue
@@ -1,6 +1,5 @@
 <template>
   <v-tooltip
-    content-class="polar-tooltip"
     :left="!isHorizontal"
     :bottom="isHorizontal"
     :disabled="hasSmallDisplay"