+++ packages/plugins/IconMenu/src/components/IconMenu.vue
@@ -15,7 +15,6 @@
       />
       <template v-else>
         <v-tooltip
-          content-class="polar-tooltip"
           :left="!isHorizontal"
           :bottom="isHorizontal"
           :disabled="hasSmallDisplay"