+++ packages/plugins/LayerChooser/src/components/LayerWrapper.vue
@@ -1,5 +1,5 @@
 <template>
-  <v-tooltip content-class="polar-tooltip" left :disabled="hasSmallDisplay">
+  <v-tooltip left :disabled="hasSmallDisplay">
     <template #activator="{ on }">
       <div
         class="polar-layer-chooser-option-line"