+++ packages/core/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## unpublished
 
-- Feature: Add globally usable css class `polar-tooltip` intended to update the appearance of `v-tooltip`s to be more visible.
+- Feature: Change the `background-color` of all `v-tooltip`s to `primary` and its `border` to `primaryContrast` to be more visible.
 - Feature: Add new state variable `hasSmallDisplay` which is updated on `resize` of the `window`.
 
 ## 1.1.0